requirements.txt
opencv-python
pillow
pynput
numpy